Arbitrary labels can be added by updating the config in the PR
There's a potential that arbitrary labels could be added to a PR by a bad actor. The mechanism to achieve this is as follows:

1. Update the auto config to have a custom label
2. Select that custom label from the PR checklist

The reason why this is problematic is that it could create a situation in which a bad actor could get code merged into a project that uses something like Peril's merge on green.

- Justin Bennett: Pull config from master branch

- Justin Bennett: Skip features if config isn't found

- Justin Bennett: Gracefully handle situation where config not found

- Justin Bennett: Ensure config return types are correct